+++ src/backend_dispatch.rs
@@ -3,7 +3,6 @@
 use crate::{backend_interface::*, backends, InitError, Rect, SoftBufferError};
 
 use raw_window_handle::{HasDisplayHandle, HasWindowHandle};
-use std::num::NonZeroU32;
 
 /// A macro for creating the enum used to statically dispatch to the platform-specific implementation.
 macro_rules! make_dispatch {
@@ -87,7 +86,7 @@ macro_rules! make_dispatch {
                 }
             }
 
-            fn resize(&mut self, width: NonZeroU32, height: NonZeroU32) -> Result<(), SoftBufferError> {
+            fn resize(&mut self, width: u32, height: u32) -> Result<(), SoftBufferError> {
                 match self {
                     $(
                         $(#[$attr])*
@@ -124,7 +123,7 @@ macro_rules! make_dispatch {
 
         impl<'a, D: HasDisplayHandle, W: HasWindowHandle> BufferInterface for BufferDispatch<'a, D, W> {
             #[inline]
-            fn width(&self) -> NonZeroU32 {
+            fn width(&self) -> u32 {
                 match self {
                     $(
                         $(#[$attr])*
@@ -134,7 +133,7 @@ macro_rules! make_dispatch {
             }
 
             #[inline]
-            fn height(&self) -> NonZeroU32 {
+            fn height(&self) -> u32 {
                 match self {
                     $(
                         $(#[$attr])*